+++ api/downlinkQueue.pb.gw.go
@@ -245,11 +245,11 @@ func RegisterDownlinkQueueHandler(ctx context.Context, mux *runtime.ServeMux, co
 }
 
 var (
-	pattern_DownlinkQueue_Enqueue_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3}, []string{"api", "nodes", "devEUI", "queue"}, ""))
+	pattern_DownlinkQueue_Enqueue_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3}, []string{"api", "devices", "devEUI", "queue"}, ""))
 
-	pattern_DownlinkQueue_Delete_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3, 1, 0, 4, 1, 5, 4}, []string{"api", "nodes", "devEUI", "queue", "id"}, ""))
+	pattern_DownlinkQueue_Delete_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3, 1, 0, 4, 1, 5, 4}, []string{"api", "devices", "devEUI", "queue", "id"}, ""))
 
-	pattern_DownlinkQueue_List_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3}, []string{"api", "nodes", "devEUI", "queue"}, ""))
+	pattern_DownlinkQueue_List_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 1, 0, 4, 1, 5, 2, 2, 3}, []string{"api", "devices", "devEUI", "queue"}, ""))
 )
 
 var (